// fe/src/lib/api/generated/model/routineIntakeStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type RoutineIntakeStatus = typeof RoutineIntakeStatus[keyof typeof RoutineIntakeStatus];


export const RoutineIntakeStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
} as const;